Switch v1.0 deprecation references
