+++ src/pytorch_lightning/overrides/base.py
@@ -75,7 +75,7 @@ class _LightningModuleWrapperBase(_DeviceDtypeModuleMixin, torch.nn.Module):
                 "`forward_module` must be a `LightningModule` instance or have an attribute `.module` pointing to one,"
                 f" got: {forward_module.__class__.__qualname__}"
             )
-        # TODO: In v1.10.0, remove the Optional type from forward_module and remove the assertion
+        # TODO: In v2.0.0, remove the Optional type from forward_module and remove the assertion
         assert forward_module is not None
         self._forward_module = forward_module
 
@@ -118,11 +118,11 @@ class _LightningModuleWrapperBase(_DeviceDtypeModuleMixin, torch.nn.Module):
         pl_module: Optional[Union["pl.LightningModule", _LightningPrecisionModuleWrapperBase]] = None,
         forward_module: Optional[Union["pl.LightningModule", _LightningPrecisionModuleWrapperBase]] = None,
     ) -> None:
-        # TODO: In v1.10.0, remove this method and mark the forward_module init argument in all subclasses as required
+        # TODO: In v2.0.0, remove this method and mark the forward_module init argument in all subclasses as required
         if pl_module is not None:
             rank_zero_deprecation(
                 f"The argument `pl_module` in `{cls.__name__}` is deprecated in v1.8.0 and will be removed in"
-                " v1.10.0. Please use `forward_module` instead."
+                " v2.0.0. Please use `forward_module` instead."
             )
         elif forward_module is None:
             raise ValueError("Argument `forward_module` is required.")
@@ -133,7 +133,7 @@ def unwrap_lightning_module(wrapped_model: nn.Module, _suppress_warning: bool = 
     attributes on the wrapper.
 
     .. deprecated:: v1.8.0
-        The function ``unwrap_lightning_module`` is deprecated in v1.8.0 and will be removed in v1.10.0. Access the
+        The function ``unwrap_lightning_module`` is deprecated in v1.8.0 and will be removed in v2.0.0. Access the
         ``LightningModule`` directly through the strategy attribute ``Strategy.lightning_module``.
 
     Raises:
@@ -142,7 +142,7 @@ def unwrap_lightning_module(wrapped_model: nn.Module, _suppress_warning: bool = 
     """
     if not _suppress_warning:
         rank_zero_deprecation(
-            "The function `unwrap_lightning_module` is deprecated in v1.8.0 and will be removed in v1.10.0. Access the"
+            "The function `unwrap_lightning_module` is deprecated in v1.8.0 and will be removed in v2.0.0. Access the"
             " `LightningModule` directly through the strategy attribute `Strategy.lightning_module`."
         )
     model = wrapped_model
