+++ src/pytorch_lightning/trainer/connectors/accelerator_connector.py
@@ -110,11 +110,11 @@ class AcceleratorConnector:
         benchmark: Optional[bool] = None,
         replace_sampler_ddp: bool = True,
         deterministic: Optional[Union[bool, _LITERAL_WARN]] = False,
-        auto_select_gpus: Optional[bool] = None,  # TODO: Remove in v1.10.0
-        num_processes: Optional[int] = None,  # deprecated
-        tpu_cores: Optional[Union[List[int], str, int]] = None,  # deprecated
-        ipus: Optional[int] = None,  # deprecated
-        gpus: Optional[Union[List[int], str, int]] = None,  # deprecated
+        auto_select_gpus: Optional[bool] = None,  # TODO: Remove in v2.0.0
+        num_processes: Optional[int] = None,  # TODO: Remove in v2.0.0
+        tpu_cores: Optional[Union[List[int], str, int]] = None,  # TODO: Remove in v2.0.0
+        ipus: Optional[int] = None,  # TODO: Remove in v2.0.0
+        gpus: Optional[Union[List[int], str, int]] = None,  # TODO: Remove in v2.0.0
     ) -> None:
         """The AcceleratorConnector parses several Trainer arguments and instantiates the Strategy including other
         components such as the Accelerator and Precision plugins.
@@ -177,9 +177,9 @@ class AcceleratorConnector:
         self._parallel_devices: List[Union[int, torch.device, str]] = []
         self._layer_sync: Optional[LayerSync] = NativeSyncBatchNorm() if sync_batchnorm else None
         self.checkpoint_io: Optional[CheckpointIO] = None
-        self._amp_type_flag: Optional[str] = None  # TODO: Remove in v1.10.0
-        self._amp_level_flag: Optional[str] = amp_level  # TODO: Remove in v1.10.0
-        self._auto_select_gpus: Optional[bool] = auto_select_gpus
+        self._amp_type_flag: Optional[str] = None  # TODO: Remove in v2.0.0
+        self._amp_level_flag: Optional[str] = amp_level  # TODO: Remove in v2.0.0
+        self._auto_select_gpus: Optional[bool] = auto_select_gpus  # TODO: Remove in v2.0.0
 
         self._check_config_and_set_final_flags(
             strategy=strategy,
@@ -380,7 +380,7 @@ class AcceleratorConnector:
         if amp_type is not None:
             rank_zero_deprecation(
                 "The NVIDIA/apex AMP implementation has been deprecated upstream. Consequently, its integration inside"
-                " PyTorch Lightning has been deprecated in v1.9.0 and will be removed in v1.10.0."
+                " PyTorch Lightning has been deprecated in v1.9.0 and will be removed in v2.0.0."
                 f" The `Trainer(amp_backend={amp_type!r})` argument is deprecated. Removing this argument will avoid"
                 f" this message, it will select PyTorch's implementation automatically."
             )
@@ -391,7 +391,7 @@ class AcceleratorConnector:
         if amp_level is not None:
             rank_zero_deprecation(
                 "The NVIDIA/apex AMP implementation has been deprecated upstream. Consequently, its integration inside"
-                " PyTorch Lightning has been deprecated in v1.9.0 and will be removed in v1.10.0."
+                " PyTorch Lightning has been deprecated in v1.9.0 and will be removed in v2.0.0."
                 f" The `Trainer(amp_level={amp_level!r})` argument is deprecated. Removing this argument will avoid"
                 f" this message."
             )
@@ -562,7 +562,7 @@ class AcceleratorConnector:
     def _set_devices_flag_if_auto_select_gpus_passed(self) -> None:
         if self._auto_select_gpus is not None:
             rank_zero_deprecation(
-                "The Trainer argument `auto_select_gpus` has been deprecated in v1.9.0 and will be removed in v1.10.0."
+                "The Trainer argument `auto_select_gpus` has been deprecated in v1.9.0 and will be removed in v2.0.0."
                 " Please use the function `pytorch_lightning.accelerators.find_usable_cuda_devices` instead."
             )
         if self._auto_select_gpus and isinstance(self._gpus, int) and isinstance(self.accelerator, CUDAAccelerator):
