+++ src/pytorch_lightning/utilities/seed.py
@@ -47,7 +47,7 @@ def isolate_rng() -> Generator[None, None, None]:
 def seed_everything(*args: Any, **kwargs: Any) -> Any:
     rank_zero_deprecation(
         "`pytorch_lightning.utilities.seed.seed_everything` has been deprecated in v1.8.0 and will be"
-        " removed in v1.10.0. Please use `lightning_fabric.utilities.seed.seed_everything` instead."
+        " removed in v2.0.0. Please use `lightning_fabric.utilities.seed.seed_everything` instead."
     )
     return new_seed_everything(*args, **kwargs)
 
@@ -55,7 +55,7 @@ def seed_everything(*args: Any, **kwargs: Any) -> Any:
 def reset_seed() -> None:
     rank_zero_deprecation(
         "`pytorch_lightning.utilities.seed.reset_seed` has been deprecated in v1.8.0 and will be"
-        " removed in v1.10.0. Please use `lightning_fabric.utilities.seed.reset_seed` instead."
+        " removed in v2.0.0. Please use `lightning_fabric.utilities.seed.reset_seed` instead."
     )
     return new_reset_seed()
 
@@ -63,6 +63,6 @@ def reset_seed() -> None:
 def pl_worker_init_function(*args: Any, **kwargs: Any) -> None:
     rank_zero_deprecation(
         "`pytorch_lightning.utilities.seed.pl_worker_init_function` has been deprecated in v1.8.0 and will be"
-        " removed in v1.10.0. Please use `lightning_fabric.utilities.seed.pl_worker_init_function` instead."
+        " removed in v2.0.0. Please use `lightning_fabric.utilities.seed.pl_worker_init_function` instead."
     )
     return new_pl_worker_init_function(*args, **kwargs)
