+++ tests/tests_pytorch/helpers/runif.py
@@ -77,8 +77,8 @@ class RunIf:
         ipu: bool = False,
         hpu: bool = False,
         mps: Optional[bool] = None,
-        horovod: bool = False,  # TODO: remove in v1.10.0
-        horovod_nccl: bool = False,  # TODO: remove in v1.10.0
+        horovod: bool = False,  # TODO: remove in v2.0.0
+        horovod_nccl: bool = False,  # TODO: remove in v2.0.0
         skip_windows: bool = False,
         standalone: bool = False,
         fairscale: bool = False,
@@ -154,7 +154,7 @@ class RunIf:
             conditions.append(not _TORCH_QUANTIZE_AVAILABLE or _miss_default)
             reasons.append("PyTorch quantization")
 
-        # TODO: remove in v1.10.0
+        # TODO: remove in v2.0.0
         if amp_apex:
             conditions.append(not _APEX_AVAILABLE)
             reasons.append("NVIDIA Apex")
